refactor: updated services
Created two separate files for page-view and post-view instead of one
